requirements.txt
Flask==2.2.2
python-telegram-bot==20.0
gunicorn==20.1.0
pyngrok==5.1.0
